作成: ChatUserListDataSource
