+++ BookList-iOS/Presentation/Presentation-Domain/Screen/Chat/ChatUserList/ChatUserListViewController.swift
@@ -6,12 +6,11 @@ final class ChatUserListViewController: UIViewController {
 
     @IBOutlet weak var tableView: UITableView!
 
-    private var users: [FirestoreUser] = []
-
     private let router: RouterProtocol = Router()
     private let disposeBag: DisposeBag = DisposeBag()
 
     private var viewModel: ChatUserListViewModel!
+    private var dataSource: ChatUserListDataSource!
 
     static func createInstance(viewModel: ChatUserListViewModel) -> ChatUserListViewController {
         let instance = ChatUserListViewController.instantiateInitialViewController()
@@ -27,9 +26,10 @@ final class ChatUserListViewController: UIViewController {
     }
 
     private func setupTableView() {
+        dataSource = ChatUserListDataSource()
         tableView.register(ChatUserListTableViewCell.xib(), forCellReuseIdentifier: ChatUserListTableViewCell.resourceName)
         tableView.tableFooterView = UIView()
-        tableView.dataSource = self
+        tableView.dataSource = dataSource
         tableView.rowHeight = 80
     }
 }
@@ -42,38 +42,10 @@ extension ChatUserListViewController {
             .drive(onNext: { [weak self] users in
                 guard let self = self else { return }
 
-                self.users = users
+                self.dataSource.chatUserList = users
                 self.tableView.reloadData()
             })
             .disposed(by: disposeBag)
-    }
-}
-
-extension ChatUserListViewController: UITableViewDataSource {
-
-    func tableView(
-        _ tableView: UITableView,
-        numberOfRowsInSection section: Int
-    ) -> Int {
-        users.count
-    }
-
-    func tableView(
-        _ tableView: UITableView,
-        cellForRowAt indexPath: IndexPath
-    ) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(
-            withIdentifier: ChatUserListTableViewCell.resourceName,
-            for: indexPath
-        )
-
-        if let chatUserListCell = cell as? ChatUserListTableViewCell {
-            if let user = users.any(at: indexPath.row) {
-                chatUserListCell.setup(user: user)
-            }
-        }
-
-        return cell
     }
 }
